refactor: 카테고리 기준 상품조회에 pagenation 추가, ai-core 모듈 삭제 및 common으로 이동
