--- ai-core/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(17)
-    }
-}
-
-dependencyManagement {
-    imports {
-        mavenBom "org.springframework.boot:spring-boot-dependencies:3.5.8"
-        mavenBom "org.springframework.ai:spring-ai-bom:1.1.2"
-    }
-}
-
-dependencies {
-    api 'io.projectreactor:reactor-core'
-    implementation 'org.springframework.boot:spring-boot-starter'
-    implementation 'org.springframework.ai:spring-ai-starter-model-openai'
-}
