+++ product/build.gradle
@@ -39,7 +39,6 @@ dependencies {
 
     runtimeOnly 'org.postgresql:postgresql'
 
-    implementation project(':ai-core')
     implementation project(':common')
 }
