+++ product/src/main/java/com/dev_high/product/domain/ProductRepository.java
@@ -22,7 +22,7 @@ public interface ProductRepository {
 
     List<Product> findByIdIn(List<String> productIds);
 
-    List<Product> findBySellerId(String sellerId);
+    List<Product> findBySellerId(String sellerId, Pageable pageable);
 
     List<Product> findByProductIds(List<String> productIds);
 }
