+++ product/src/main/java/com/dev_high/product/infraStructure/ProductJpaRepository.java
@@ -19,5 +19,5 @@ public interface ProductJpaRepository extends JpaRepository<Product, String> {
 
     List<Product> findByIdIn(List<String> ids);
 
-    List<Product> findBySellerIdAndDeletedYn(String sellerId, DeleteStatus deletedYn);
+    List<Product> findBySellerIdAndDeletedYn(String sellerId, DeleteStatus deletedYn, Pageable pageable);
 }
