+++ product/src/main/java/com/dev_high/product/infraStructure/ProductRepositoryAdapter.java
@@ -32,8 +32,8 @@ public class ProductRepositoryAdapter implements ProductRepository {
     }
 
     @Override
-    public List<Product> findBySellerId(String sellerId) {
-        return productJpaRepository.findBySellerIdAndDeletedYn(sellerId, Product.DeleteStatus.N);
+    public List<Product> findBySellerId(String sellerId, Pageable pageable) {
+        return productJpaRepository.findBySellerIdAndDeletedYn(sellerId, Product.DeleteStatus.N, pageable);
     }
 
     @Override
