+++ settings.gradle
@@ -3,7 +3,6 @@ rootProject.name = 'more'
 
 // infrastructure & common module
 include 'common'
-include 'ai-core'
 include 'config'
 include 'apigateway'
 include 'discovery'
